PLAT-781: fix cycled imports
