+++ ledger-core/configuration/configuration.go
@@ -6,14 +6,10 @@
 package configuration
 
 import (
-	"crypto"
 	"fmt"
 
 	"github.com/insolar/insconfig"
 	"gopkg.in/yaml.v2"
-
-	"github.com/insolar/assured-ledger/ledger-core/cryptography"
-	"github.com/insolar/assured-ledger/ledger-core/network/mandates"
 )
 
 const InsolarEnvPrefix = "insolar"
@@ -60,16 +56,6 @@ type BaseCloudConfig struct {
 	Log                 Log
 	PulsarConfiguration PulsarConfiguration
 	NodeConfigPaths     []string
-}
-
-type CertManagerFactory func(publicKey crypto.PublicKey, keyProcessor cryptography.KeyProcessor, certPath string) (*mandates.CertificateManager, error)
-type KeyStoreFactory func(path string) (cryptography.KeyStore, error)
-
-type CloudConfigurationProvider struct {
-	CloudConfig        BaseCloudConfig
-	CertificateFactory CertManagerFactory
-	KeyFactory         KeyStoreFactory
-	GetAppConfigs      func() []Configuration
 }
 
 // Configuration contains configuration params for all Insolar components
