+++ ledger-core/instrumentation/insapp/server_multi.go
@@ -7,16 +7,29 @@ package insapp
 
 import (
 	"context"
+	"crypto"
 
 	"github.com/insolar/component-manager"
 
 	"github.com/insolar/assured-ledger/ledger-core/appctl/beat"
 	"github.com/insolar/assured-ledger/ledger-core/configuration"
+	"github.com/insolar/assured-ledger/ledger-core/cryptography"
 	"github.com/insolar/assured-ledger/ledger-core/network"
+	"github.com/insolar/assured-ledger/ledger-core/network/mandates"
 	"github.com/insolar/assured-ledger/ledger-core/network/messagesender"
 	"github.com/insolar/assured-ledger/ledger-core/network/nodeinfo"
 	"github.com/insolar/assured-ledger/ledger-core/vanilla/throw"
 )
+
+type CertManagerFactory func(publicKey crypto.PublicKey, keyProcessor cryptography.KeyProcessor, certPath string) (*mandates.CertificateManager, error)
+type KeyStoreFactory func(path string) (cryptography.KeyStore, error)
+
+type CloudConfigurationProvider struct {
+	CloudConfig        configuration.BaseCloudConfig
+	CertificateFactory CertManagerFactory
+	KeyFactory         KeyStoreFactory
+	GetAppConfigs      func() []configuration.Configuration
+}
 
 // NetworkSupport provides network-related functions to an app compartment
 type NetworkSupport interface {
@@ -39,7 +52,7 @@ type NetworkInitFunc = func(configuration.Configuration, *component.Manager) (Ne
 // to initialize instantiate a network support for each app compartment (one per node). A default implementation is applied when NetworkInitFunc is nil.
 type MultiNodeConfigFunc = func(baseCfg configuration.Configuration) ([]configuration.Configuration, NetworkInitFunc)
 
-func NewMulti(cfg configuration.CloudConfigurationProvider, appFn AppFactoryFunc, multiFn MultiNodeConfigFunc, extraComponents ...interface{}) *Server {
+func NewMulti(cfg CloudConfigurationProvider, appFn AppFactoryFunc, multiFn MultiNodeConfigFunc, extraComponents ...interface{}) *Server {
 	if multiFn == nil {
 		panic(throw.IllegalValue())
 	}
